requirements.txt
Django==1.7.1
Fabric==1.10.1
Markdown==2.5.2
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.19
celery==3.1.17
cloudinary==1.0.19
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.9.1
django-toolbelt==0.0.1
djangorestframework==3.0.2
ecdsa==0.11
gunicorn==19.1.1
kombu==3.0.24
paramiko==1.15.2
psycopg2==2.5.4
pycrypto==2.6.1
pymongo==2.7.2
pytz==2014.10
static3==0.5.1
wsgiref==0.1.2
